try a simple analysis

+++ benchmarks/diffractive_vm/analysis/diffractive_vm_simple_analysis.cxx
@@ -60,7 +60,7 @@ int diffractive_vm_simple_analysis(const std::string& config_name)
     TTreeReaderArray<float> reco_px_array = {tree_reader, "ReconstructedChargedParticles.momentum.x"};
     TTreeReaderArray<float> reco_py_array = {tree_reader, "ReconstructedChargedParticles.momentum.y"};
     TTreeReaderArray<float> reco_pz_array = {tree_reader, "ReconstructedChargedParticles.momentum.z"};
-    TTreeReaderArray<int> reco_charge_array = {tree_reader, "ReconstructedChargedParticles.charge"};
+    TTreeReaderArray<float> reco_charge_array = {tree_reader, "ReconstructedChargedParticles.charge"};
 
     TTreeReaderArray<unsigned int> rec_id = {tree_reader, "ReconstructedChargedParticlesAssociations.recID"};
     TTreeReaderArray<unsigned int> sim_id = {tree_reader, "ReconstructedChargedParticlesAssociations.simID"};
